requirements.txt
# Semantic Opinion Dynamics - Dependencies
networkx>=3.2
numpy>=1.24
sentence-transformers>=2.2
torch>=2.0
google-generativeai>=0.3
python-dotenv>=1.0
matplotlib>=3.7
scipy>=1.11
